+++ CHANGELOG.md
@@ -5,6 +5,7 @@ API Changes:
 * auto_fragment is a member of stream
 * binary, text are members of stream
 * read_buffer_size is a member of stream
+* read_message_max is a member of stream
 
 Actions Required:
 
@@ -16,6 +17,9 @@ Actions Required:
 
 * Change call sites which use read_buffer_size with set_option to
   call stream::read_buffer_size instead.
+
+* Change call sites which use read_message_max with set_option to
+  call stream::read_message_max instead.
 
 --------------------------------------------------------------------------------